test(markdown_parser): move weight tests to separate file
Repo convention places unit tests in a sibling ${filename}_tests.rs
file, not an inline #[cfg(test)] module. Follow the pattern already
used by html_parser.rs / html_parser_tests.rs in the same crate.
